Reorganize addInnerHits
