+++ lib/elasticsearch/elastic-query-builder.js
@@ -11,9 +11,9 @@ const POPULARITY_BOOSTS = [
 ]
 
 class ElasticQueryBuilder {
-  constructor (apiRequest) {
+  constructor (apiRequest, options = {}) {
     this.request = apiRequest
-    this.query = new ElasticQuery()
+    this.query = new ElasticQuery(options)
 
     // Break on search_scope:
     switch (this.request.params.search_scope) {
@@ -706,8 +706,8 @@ class ElasticQueryBuilder {
   /**
   * Create a ElasticQueryBuilder for given ApiRequest instance
   */
-  static forApiRequest (request) {
-    return new ElasticQueryBuilder(request)
+  static forApiRequest (request, options = {}) {
+    return new ElasticQueryBuilder(request, options)
   }
 }
