+++ lib/elasticsearch/elastic-query.js
@@ -7,10 +7,11 @@
 **/
 
 class ElasticQuery {
-  constructor () {
+  constructor (options = {}) {
     this.musts = []
     this.shoulds = []
     this.filters = []
+    this.options = options
   }
 
   addMust (clause) {
@@ -42,7 +43,8 @@ class ElasticQuery {
   * "query" param in a ES call
   */
   toJson () {
-    if (!this.musts.length && !this.shoulds.length && !this.filters.length) {
+    console.log('options: ', this.options)
+    if (!this.musts.length && !this.shoulds.length && !this.filters.length && !this.options.items) {
       return {
         match_all: {}
       }
@@ -56,7 +58,7 @@ class ElasticQuery {
     if (this.shoulds.length) {
       result.bool.should = this.shoulds
     }
-    if (this.filters.length) {
+    if (this.filters.length || this.options.items) {
       result.bool.filter = this.filters
     }
